% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blsdata6.R
\docType{data}
\name{blsdata6}
\alias{blsdata6}
\title{Box Lunch Study - Baseline dataset}
\format{
A data frame with 156 rows and 6 variables
}
\usage{
data(blsdata6)
}
\description{
The variables are as follows:
}
\details{
\itemize{
\item trt. Treatment
\item bmi0. BMI
\item kcal24h6.
\item edeq14.0
\item age. Age
\item tfactor3.0
}
}
\examples{
data(blsdata6)

}
\keyword{datasets}
