% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_func.R
\name{est_fncn}
\alias{est_fncn}
\title{Estimation function over the test sample to determine the CATE (double-sample tree)}
\usage{
est_fncn(Y, Z, X, X.i, grp.ids, LT)
}
\arguments{
\item{Y}{Response vector}

\item{Z}{Treatment indicator}

\item{X}{Covariate dataset}

\item{X.i}{Covariate dataset corresponding to a test sample}

\item{grp.ids}{Determine the index values relevant to the identified subgroups}

\item{LT}{Tree structure}
}
\description{
Estimation function over the test sample to determine the CATE (double-sample tree)
}
