% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_func.R
\name{grp_fnnc}
\alias{grp_fnnc}
\title{Get the ids from the derived tree structure (double-sample tree)}
\usage{
grp_fnnc(LT, id)
}
\arguments{
\item{LT}{Tree structure built over the training sample}

\item{id}{Relevant index}
}
\description{
Get the ids from the derived tree structure (double-sample tree)
}
