% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_func.R
\name{grp_func}
\alias{grp_func}
\title{Group ID Function - Double Sample Approach
Implementation described in est_fncn()}
\usage{
grp_func(LT, id)
}
\arguments{
\item{LT}{Tree List}

\item{id}{ID}
}
\description{
Group ID Function - Double Sample Approach
Implementation described in est_fncn()
}
\keyword{Subgroup}
\keyword{Tree}
