% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMEstree.R
\name{LMEstree}
\alias{LMEstree}
\title{Build a single-sample tree}
\usage{
LMEstree(Y, X, ctl.ind, indx, pval.thresh = 0.2, min.leaf.size = 12,
  min.split.size = 7, split.pts = 0.05, min.eff.size = 0.25 * sd(Y,
  na.rm = T))
}
\arguments{
\item{Y}{Response vector}

\item{X}{Covariate matrix}

\item{ctl.ind}{Index corresponding to the control subject}

\item{indx}{Index to be selected for use in the tree-building process}

\item{pval.thresh}{Threshold of selected p-value}

\item{min.leaf.size}{Minium size of a terminal node}

\item{min.split.size}{Minimum size of a node to split}

\item{split.pts}{Set the increment in a sequence to search for split points}
}
\value{
A list object
}
\description{
Build a single-sample tree
}
