% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMEtree.R
\name{LMEtree}
\alias{LMEtree}
\title{Build the double-sample tree structure and specify the appropriate parameters}
\usage{
LMEtree(Y, X, ctl.ind, indx, indx.TEST, X.TEST, pval.thresh = 0.05,
  min.leaf.size = 20, min.split.size = 10, split.pts = 0.05)
}
\arguments{
\item{Y}{Response vector}

\item{X}{Covariates matrix}

\item{ctl.ind}{Control Index}

\item{indx}{Index}

\item{indx.TEST}{Index in the Test dataset}

\item{X.TEST}{The defined covariate test dataset}

\item{pval.thresh}{p-value threshold}

\item{min.leaf.size}{Minimum lead size}

\item{min.split.size}{Minimum split size}

\item{split.pts}{Increment of the sequence to define split points}
}
\description{
Build the double-sample tree structure and specify the appropriate parameters
}
