% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTfunction.R
\name{LTfunction}
\alias{LTfunction}
\title{Use the matched dataset for the LMEdtree (double-sample tree)}
\usage{
LTfunction(Y, Z, X, X.i)
}
\arguments{
\item{Y}{Response vector}

\item{Z}{Treatment indicator}

\item{X}{Covariate matrix}

\item{X.i}{Covariate matrix corresponding to the test sample}
}
\description{
Use the matched dataset for the LMEdtree (double-sample tree)
}
