% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatchForTree.R
\name{MatchForTree}
\alias{MatchForTree}
\title{Matching Function}
\usage{
MatchForTree(Y, Z, X, version = "standard", nmatch = 1)
}
\arguments{
\item{Y}{Response vector}

\item{Z}{Treatment indicator}

\item{X}{Covariate matrix}

\item{version}{Version of the code}

\item{nmatch}{Number matched}
}
\value{
A list object that records the response, covariate, treated index and control index from the matched dataset
}
\description{
Matching Function
}
\keyword{matching}
