% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeMat.R
\name{TreeMat}
\alias{TreeMat}
\title{Use the constructed single-sample tree and estimate the effects over the same sample}
\usage{
TreeMat(orig.LT, orig.Y, orig.X)
}
\arguments{
\item{orig.LT}{Use the constructed tree structure}

\item{orig.Y}{Use the response vector (original, not matched)}

\item{orig.X}{Use the covariate matrix (original, not matched)}
}
\value{
A matrix composed of the tree structure and corresponding effects
}
\description{
Use the constructed single-sample tree and estimate the effects over the same sample
}
