% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTfunction.R
\name{LTfunction}
\alias{LTfunction}
\title{Determines the splitting process and split criterions}
\usage{
LTfunction(
  Y,
  Z,
  X,
  X.i,
  version = "prognostic",
  pval.thresh = 0.05,
  min.split.size = 10,
  sample = "single"
)
}
\arguments{
\item{Y}{response vector}

\item{Z}{treatment indicator}

\item{X}{covariate matrix}

\item{X.i}{covariate matrix - ids}

\item{version}{type of score - default is the prognostic score}

\item{pval.thresh}{Threshold of p-value}

\item{min.split.size}{Minimum size to split}

\item{sample}{Use a single or double sample - default is single. The double sample approach uses a separate sample to construct the tree and to estimate the effects. This can be modified as required.}
}
\description{
This function generates the tree structure as a list function.
}
\examples{
set.seed(10)
N<-2000
numx <- 5
alpha <-0.8
theta<-0.8
beta<- c(1,.8,.6,.4,.2)
gamma <- 1
Z <- rep(c(0, 1), each = N/2)
sigma <- diag(numx)
X.i <- mvrnorm(N,mu=rep(0,numx),Sigma=sigma)
W <- Z * ifelse(X.i[,1] > 0, 1, 0)
mu <- alpha + theta*Z + X.i \%*\% beta + W * gamma
Y <- rnorm(N, mean=mu)

# Single sample approach
LT1<-LTfunction(Y, as.logical(Z), X.i, X.i)

# Double sample approach
Nused <- 200
subjects <- c(1:(Nused/2), 1001:(1000+Nused/2))
LT2<-LTfunction(Y[subjects], Z[subjects], X.i[subjects,], X.i[subjects,], sample = "double")


}
