% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatchForTree.R
\name{MatchForTree}
\alias{MatchForTree}
\title{Function to construct matched pairs over treated and control subjects and compute their relevant causal effect
(Relevant for both the double-sample and single-sample tree)}
\usage{
MatchForTree(Y, Z, X, version = "standard")
}
\arguments{
\item{Y}{Response vector}

\item{Z}{Treatment indicator}

\item{X}{Covariate matrix}

\item{version}{Type of matching (i.e., standard, prognostic, Xmin, Xtrt, biasadjusted, autobalance)}
}
\value{
A list object that records the response, covariate, treated index and control index from the matched dataset
}
\description{
Function to construct matched pairs over treated and control subjects and compute their relevant causal effect
(Relevant for both the double-sample and single-sample tree)
}
\keyword{matching}
