% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatchForTree.R
\name{MatchForTree}
\alias{MatchForTree}
\title{Matching Function}
\usage{
MatchForTree(Y, Z, X, version = "prognostic")
}
\arguments{
\item{Y}{response vector}

\item{Z}{treatment indicator}

\item{X}{covariate matrix}

\item{version}{version type of matching to be performed}
}
\description{
This function is used for matching with replacement and determines the treatment effect and relevant covariates for the matched pairs.
}
\examples{
N<-2000
numx <- 5
alpha <-0.8
theta<-0.8
beta<- c(1,.8,.6,.4,.2)
gamma <- 1
Z <- rep(c(0, 1), each = N/2)
sigma <- diag(numx)
X.i <- mvrnorm(N,mu=rep(0,numx),Sigma=sigma)
W <- Z * ifelse(X.i[,1] > 0, 1, 0)
mu <- alpha + theta*Z + X.i \%*\% beta + W * gamma
Y <- rnorm(N, mean=mu)

MatchForTree(Y, Z, X.i)
}
\keyword{matching}
\keyword{prognostic}
\keyword{treatment}
