% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeMat.R
\name{TreeMat}
\alias{TreeMat}
\title{TEHTree matrix}
\usage{
TreeMat(orig.LT, orig.Y, orig.X)
}
\arguments{
\item{orig.LT}{Output from the LMEtree}

\item{orig.Y}{Responses from the matched dataset}

\item{orig.X}{Covariates from the matched dataset}
}
\value{
Tree structure recorded as a matrix
}
\description{
TEHTree matrix
}
